First commit
This project really should be integrated with necrobot since it's really
just some modules building off of that, and I'm planning on reusing a
ton of the code. That said, I'm on a deadline and it's easier at the
moment to just work on it as its own thing and copy relevant stuff back
and forth. However, integrating the two in a good way is a future TODO
if possible (so that bugfixes on either affect the other).